requirements.txt
# Core
python-dotenv
langchain-core
langchain-openai
langchain-anthropic
langchain-text-splitters
langgraph

# Document loading
pdfplumber

# RAG / Vector
pgvector
psycopg[binary]
qdrant-client

# Session / Rate limiting
redis

# Observability
prometheus_client

# Testing
pytest
pytest-asyncio
pytest-mock
pytest-cov

# Evaluation
ragas
datasets

# Web / Scraping (for research agent)
playwright
aiofiles
beautifulsoup4
